Facade layer, interfaces added
