+++ src/main/java/com/demoproject/cricketapp/beans/request/MatchRequest.java
@@ -1,0 +1,14 @@
+package com.demoproject.cricketapp.beans.request;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class MatchRequest {
+
+    private LocalDateTime dateTime;
+    private String team1Id;
+    private String team2Id;
+    private int overs;
+}
